homework_07/mysite/requirements.txt
asgiref==3.3.1
certifi==2021.10.8
charset-normalizer==2.0.8
Django==3.1.4
djangorestframework==3.12.4
gunicorn==20.0.4
idna==3.3
lxml==4.6.2
psycopg2-binary==2.9.2
pydantic==1.8.2
python-dotenv==0.19.2
pytz==2020.5
requests==2.26.0
sqlparse==0.4.1
typing_extensions==4.0.0
urllib3==1.26.7
